+++ src/columns/esp.rs
@@ -28,7 +28,7 @@ impl Esp {
 impl Column for Esp {
     fn add(&mut self, proc: &ProcessInfo) {
         let raw_content = proc.curr_proc.stat.kstkesp;
-        let fmt_content = format!("{:x}", raw_content);
+        let fmt_content = format!("{:016x}", raw_content);
 
         self.fmt_contents.insert(proc.pid, fmt_content);
         self.raw_contents.insert(proc.pid, raw_content);